# src/main/java/com/example/bankcards/dto/auth/AuthResponse.java
package com.example.bankcards.dto.auth;

public record AuthResponse(String token) {
}
//	Возвращаем только JWT — фронту достаточно положить его в Authorization: Bearer