# src/main/java/com/example/bankcards/dto/auth/LoginRequest.java
package com.example.bankcards.dto.auth;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(@NotBlank String username,
                           @NotBlank String password
) {}